refactor: changed name of package
