+++ app/src/main/java/com/liwid/liwid_extension/example/model/MatchDataModel.kt
@@ -1,4 +1,4 @@
-package com.example.liwid_app.example.model
+package com.liwid.liwid_extension.example.model
 
 import com.google.gson.annotations.SerializedName
