+++ app/src/main/java/com/liwid/liwid_extension/extension/widget/LiveSportsWidget.kt
@@ -1,11 +1,11 @@
-package com.example.liwid_app.extension.widget
+package com.liwid.liwid_extension.extension.widget
 import android.content.Context
 import android.app.Activity
 import android.util.Log
-import com.example.liwid_app.extension.LiveWidget
-import com.example.liwid_app.extension.api.ApiClient
-import com.example.liwid_app.extension.api.model.SportsData
-import com.example.liwid_app.extension.api.model.SportsDataResponse
+import com.liwid.liwid_extension.extension.LiveWidget
+import com.liwid.liwid_extension.extension.api.ApiClient
+import com.liwid.liwid_extension.extension.api.model.SportsData
+import com.liwid.liwid_extension.extension.api.model.SportsDataResponse
 import retrofit2.Call
